+++ msa/black-box-tests/src/test/java/org/thingsboard/server/msa/connectivity/CoapClientTest.java
@@ -16,7 +16,6 @@
 package org.thingsboard.server.msa.connectivity;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.gson.JsonObject;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
@@ -26,18 +25,14 @@ import org.thingsboard.server.common.data.Device;
 import org.thingsboard.server.common.data.DeviceProfile;
 import org.thingsboard.server.common.data.DeviceProfileProvisionType;
 import org.thingsboard.server.common.data.security.DeviceCredentials;
-import org.thingsboard.server.common.msg.session.FeatureType;
-import org.thingsboard.server.msa.AbstractContainerTest;
+import org.thingsboard.server.msa.AbstractCoapClientTest;
 import org.thingsboard.server.msa.DisableUIListeners;
-import org.thingsboard.server.msa.TestCoapClient;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.thingsboard.server.msa.prototypes.DevicePrototypes.defaultDevicePrototype;
 
 @DisableUIListeners
-public class CoapClientTest  extends AbstractContainerTest {
-    private TestCoapClient client;
-
+public class CoapClientTest extends AbstractCoapClientTest{
     private Device device;
     @BeforeMethod
     public void setUp() throws Exception {
@@ -48,6 +43,7 @@ public class CoapClientTest  extends AbstractContainerTest {
     @AfterMethod
     public void tearDown() {
         testRestClient.deleteDeviceIfExists(device.getId());
+        disconnect();
     }
 
     @Test
@@ -101,21 +97,5 @@ public class CoapClientTest  extends AbstractContainerTest {
 
         assertThat(response.get("status").asText()).isEqualTo("NOT_FOUND");
     }
-
-    private byte[] createCoapClientAndPublish(String deviceName) throws Exception {
-        String provisionRequestMsg = createTestProvisionMessage(deviceName);
-        client = new TestCoapClient(TestCoapClient.getFeatureTokenUrl(FeatureType.PROVISION));
-        return client.postMethod(provisionRequestMsg.getBytes()).getPayload();
-    }
-
-    private String createTestProvisionMessage(String deviceName) {
-        ObjectNode provisionRequest = JacksonUtil.newObjectNode();
-        provisionRequest.put("provisionDeviceKey", TEST_PROVISION_DEVICE_KEY);
-        provisionRequest.put("provisionDeviceSecret", TEST_PROVISION_DEVICE_SECRET);
-        if (deviceName != null) {
-            provisionRequest.put("deviceName", deviceName);
-        }
-        return provisionRequest.toString();
-    }
-
 }
+
